Matthew Lee sprint 1 backend work
Used a skeleton frame for html and set up just a basic working of the typescript I/O. Still need to develop an effective way to store all the information as well. Also slight confusion as I was planning on developing the account setup backend work and the account login front end work has been completed. Not a problem just need to make sure for next sprint that I develop backend work for the login aspect of the site. Also not sure if pushing this code onto the main branch will conflict with the front end work that has already been completed but if it does I can remove this, just wanted to push some code out so we had a starting frame to work with.